refactor: unify tab bar style, merge Tools into Settings, rename to Automation
- Tab bars moved from PageHeader right slot to underline style below
  header (consistent with DevPage) across Automation, Logs, News, Settings
- Scheduler renamed to Automation (broader than just scheduling)
- Tools merged into Settings as a tab (agent implementation detail)
- Sidebar reorganized: new Integration group (Connectors, Trading Account,
  AI Provider), Logs moved to System, Tools entry removed
- Old URLs (/scheduler, /tools) redirect to new locations
