+++ ui/src/components/Sidebar.tsx
@@ -52,21 +52,11 @@ const NAV_SECTIONS: NavSection[] = [
     sectionLabel: 'Agent',
     items: [
       {
-        page: 'scheduler',
-        label: 'Scheduler',
+        page: 'automation',
+        label: 'Automation',
         icon: (active) => (
           <svg width="18" height="18" viewBox="0 0 24 24" fill={active ? 'currentColor' : 'none'} stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
-            <circle cx="12" cy="12" r="10" />
-            <polyline points="12 6 12 12 16 14" />
-          </svg>
-        ),
-      },
-      {
-        page: 'logs',
-        label: 'Logs',
-        icon: (active) => (
-          <svg width="18" height="18" viewBox="0 0 24 24" fill={active ? 'currentColor' : 'none'} stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
-            <polyline points="22 12 18 12 15 21 9 3 6 12 2 12" />
+            <path d="M13 2L3 14h9l-1 8 10-12h-9l1-8z" />
           </svg>
         ),
       },
@@ -101,7 +91,7 @@ const NAV_SECTIONS: NavSection[] = [
     ],
   },
   {
-    sectionLabel: 'System',
+    sectionLabel: 'Integration',
     items: [
       {
         page: 'connectors',
@@ -115,17 +105,8 @@ const NAV_SECTIONS: NavSection[] = [
         ),
       },
       {
-        page: 'tools',
-        label: 'Tools',
-        icon: (active) => (
-          <svg width="18" height="18" viewBox="0 0 24 24" fill={active ? 'currentColor' : 'none'} stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
-            <path d="M14.7 6.3a1 1 0 0 0 0 1.4l1.6 1.6a1 1 0 0 0 1.4 0l3.77-3.77a6 6 0 0 1-7.94 7.94l-6.91 6.91a2.12 2.12 0 0 1-3-3l6.91-6.91a6 6 0 0 1 7.94-7.94l-3.76 3.76z" />
-          </svg>
-        ),
-      },
-      {
         page: 'trading' as const,
-        label: 'Trading',
+        label: 'Trading Account',
         icon: (active: boolean) => (
           <svg width="18" height="18" viewBox="0 0 24 24" fill={active ? 'currentColor' : 'none'} stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
             <path d="M2 20h20" />
@@ -143,6 +124,20 @@ const NAV_SECTIONS: NavSection[] = [
           <svg width="18" height="18" viewBox="0 0 24 24" fill={active ? 'currentColor' : 'none'} stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
             <path d="M12 2a2 2 0 0 1 2 2c0 .74-.4 1.39-1 1.73v1.27h1a7 7 0 0 1 7 7h1.27c.34-.6.99-1 1.73-1a2 2 0 1 1 0 4c-.74 0-1.39-.4-1.73-1H21a7 7 0 0 1-7 7v1.27c.6.34 1 .99 1 1.73a2 2 0 1 1-4 0c0-.74.4-1.39 1-1.73V21a7 7 0 0 1-7-7H2.73c-.34.6-.99 1-1.73 1a2 2 0 1 1 0-4c.74 0 1.39.4 1.73 1H4a7 7 0 0 1 7-7V5.73c-.6-.34-1-.99-1-1.73a2 2 0 0 1 2-2z" />
             <circle cx="12" cy="14" r="3" />
+          </svg>
+        ),
+      },
+    ],
+  },
+  {
+    sectionLabel: 'System',
+    items: [
+      {
+        page: 'logs',
+        label: 'Logs',
+        icon: (active) => (
+          <svg width="18" height="18" viewBox="0 0 24 24" fill={active ? 'currentColor' : 'none'} stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
+            <polyline points="22 12 18 12 15 21 9 3 6 12 2 12" />
           </svg>
         ),
       },
